feat(pipelines): produce float32 normalized data in CCD path (#147)
The CCD pipelines computed the entire normalization in float64 and only
downcast to float32 on write, doubling the in-memory footprint of large
image stacks for no benefit -- neutron detectors are 16-bit and float32
(~7 significant figures) is sufficient. The event/TOF path was already
float32, so this was a CCD-only asymmetry.

- loaders/tiff_loader.py, loaders/fits_loader.py: load image data as
  float32 (variances inherit float32), so the whole CCD path stays float32.
- pipelines/venus_ccd.py: cast the IntegratedPCharge coords to float32
  before normalize_transmission. The coord is float64 (parsed via float()),
  and scipp promotes float32/float64 -> float64, which would otherwise
  silently re-promote VENUS back to float64 at the proton-charge division.
- pipelines/venus_ccd.py, pipelines/mars_ccd.py: cast the final transmission
  to float32 before write/return -- a robust, symmetric guarantee of the
  float32 product regardless of intermediate dtype promotion.

UQ is unchanged: scipp variances propagate identically, just in float32;
no term is added or dropped. float32 matches float64 to ~7 sig figs in
values and variances (verified). On-disk HDF5/TIFF output was already
float32, so file dtypes are unchanged; written values may differ by up to
~1e-7 (now computed in float32 rather than rounded from float64).

Tests: loader-contract float32 assertions; pipeline-returns-float32 for
VENUS (with-dark and no-dark) and MARS; the no-dark UQ structural oracle
now mirrors the pipeline's float32 handling (the pinned-constant oracle at
rtol=1e-3 remains the independent first-principles guard). Pinned-value
transmission assertions now use an explicit rtol=1e-5 appropriate to float32
precision rather than the default 1e-7 (per AGENTS.md). Full suite green,
pre-commit clean.

Reviewed by a dual-LLM-family (Claude + Codex) detection-only pass: 0 P0/P1.

Assisted-With: Claude Opus 4.8 <noreply@anthropic.com>

+++ src/neunorm/loaders/fits_loader.py
@@ -57,8 +57,10 @@ def load_fits_stack(paths: Sequence[str | Path], tof_edges: Optional[np.ndarray]
                 hdul.info(output=info_buf)
                 logger.debug("FITS info for {}:\n{}", path, info_buf.getvalue().rstrip())
 
-                # Assume data is in primary HDU
-                arr = hdul[0].data.astype(np.float64)
+                # Assume data is in primary HDU. float32 is sufficient for neutron
+                # imaging (16-bit detectors) and halves the in-memory footprint of
+                # large stacks (issue #147).
+                arr = hdul[0].data.astype(np.float32)
                 data_list.append(arr)
 
                 # Store header from first file
